Add v2 of the server!!
This server now receives data from a ecowitt gw1000 gateway which allows for
getting of all temps and humidities from your ecowitt devices. This means that
the client is no longer needed for this. Additionally a lot of changes were made
to support a 24 hour schedule of the system, saving of statistics, and AP setup mode.